* feature/Checkout_API:
  Feat: use CheckoutRequest to prepare the cart
  Fix: Int instead of double for Drink id
  Feat: implement Drink serialize method
  Feat: implementation
  Fix: test not working
  Feat: serialization test written
  Feat: declare pizza serialization test
  Feat: define base initialization pizza tests
  Feat: add serialization method on ShippableItem
  Feat: define base class with properties
  Use generated url
  Feat: move Cart insite the Checkout group
  Feat: define Checkout request
  Feat: Move NSError base errors into a separate extension
  Feat: check url composition
  Feat: first test passed
  Feat: add first test who fails
  Declare CheckoutManager
  Feat: add test Target

+++ Nennos Pizza/Model/API/API Models/Drink.swift
@@ -10,13 +10,13 @@ import Foundation
 
 struct Drink: PricedItem {
     
-    let id: Double
+    let id: Int
     let name: String
     let price: Double
     
      init?(resource: Dictionary<String, Any>) {
         
-        guard let id = resource["id"] as? Double,
+        guard let id = resource["id"] as? Int,
             let name = resource["name"] as? String,
             let price = resource["price"] as? Double else {
